chore: extract tiles from maps
